+++ bauh/view/qt/colors.py
@@ -1,3 +1,0 @@
-GREEN = '#91a069'
-ORANGE = '#ECB03E'
-RED = '#FF614E'